+++ packages/discord.js-utilities/src/lib/MessagePrompter/MessagePrompter.ts
@@ -1,4 +1,4 @@
-import type { Ctor } from '@sapphire/utilities/utilityTypes';
+import type { Ctor } from '@sapphire/utilities/types';
 import type { CollectorFilter, EmojiResolvable, Message, MessageReaction, User } from 'discord.js';
 import type { MessagePrompterChannelTypes, MessagePrompterMessage } from './constants';
 import type {
