+++ packages/discord.js-utilities/src/lib/MessagePrompter/strategies/MessagePrompterBaseStrategy.ts
@@ -1,4 +1,4 @@
-import type { Awaitable } from '@sapphire/utilities/utilityTypes';
+import type { Awaitable } from '@sapphire/utilities/types';
 import { isNullish } from '@sapphire/utilities/isNullish';
 import type { CollectorFilter, CollectorOptions, EmojiIdentifierResolvable, Message, MessageReaction, User } from 'discord.js';
 import { isTextBasedChannel } from '../../type-guards';
