+++ packages/utilities/src/index.ts
@@ -39,4 +39,4 @@ export * from './lib/sleep';
 export * from './lib/splitText';
 export { toTitleCase, type ToTitleCaseOptions } from './lib/toTitleCase';
 export * from './lib/tryParse';
-export * from './lib/utilityTypes';
+export * from './lib/types';
