+++ packages/utilities/src/lib/filterNullAndUndefined.ts
@@ -1,5 +1,5 @@
 import { isNullish } from './isNullOrUndefined';
-import type { Nullish } from './utilityTypes';
+import type { Nullish } from './types';
 
 /**
  * Checks whether a value is not `null` nor `undefined`.
