+++ packages/utilities/src/lib/isNullOrUndefined.ts
@@ -1,4 +1,4 @@
-import type { Nullish } from './utilityTypes';
+import type { Nullish } from './types';
 
 /**
  * Checks whether or not a value is `null` or `undefined`
