+++ packages/utilities/src/lib/isNullOrUndefinedOrZero.ts
@@ -1,5 +1,5 @@
 import { isNullish } from './isNullOrUndefined';
-import type { Nullish } from './utilityTypes';
+import type { Nullish } from './types';
 
 /**
  * Checks whether or not a value is `null`, `undefined` or `0`
